fix(java_tls): kd_i64 signed-shift UB + add offline kafka_parser tests
kafka_parser.h's kd_i64 assembled 8 bytes into a *signed* int64_t with
`v = (v << 8) | byte`. Once the value's sign bit is set, that signed left
shift is undefined behavior (UBSan: "left shift of N by 8 places cannot be
represented in type 'long int'"). Accumulate into uint64_t, cast once at the
end. kd_i16/kd_i32 already use unsigned casts and are unaffected.

Add self-contained tests (no live Kafka, no cluster):
- tests/test_kafka_parser.c: round-trips gzip/zstd/lz4 (multi-block) and
  snappy (raw + multi-block xerial) through kafka_decompress, and parses a
  full Produce v9 (flexible) frame down to records (29 checks).
- tests/fuzz_kafka_parser.c: ASan+UBSan fuzz over all parser entry points
  (truncation / byte-flip / random / ~2GB-block xerial) — clean after the
  fix; this harness is what caught the kd_i64 UB.
- tests/run.sh + README.

+++ src/stirling/source_connectors/socket_tracer/java_tls/collector/kafka_parser.h
@@ -129,9 +129,9 @@ static int32_t kd_i32(kd_t *d) {
 }
 static int64_t kd_i64(kd_t *d) {
   if (!kd_need(d, 8)) return 0;
-  int64_t v = 0;
+  uint64_t v = 0;  /* accumulate unsigned: a signed left-shift that sets the sign bit is UB */
   for (int i = 0; i < 8; i++) v = (v << 8) | d->p[i];
-  d->p += 8; return v;
+  d->p += 8; return (int64_t)v;
 }
 /* unsigned varint (LEB128), up to 5 bytes (35 bits like Pixie). */
 static uint32_t kd_uvarint(kd_t *d) {
